fix: corrigir status de trilhas concluídas (float vs int)
O round() do PHP retorna float, causando falha na comparação estrita
=== 100. Agora progress_percent é castado para int no controller e a
view usa >= 100 para determinar conclusão.

+++ app/Http/Controllers/Employee/PathController.php
@@ -25,7 +25,7 @@ class PathController extends Controller
                     ->count();
                 $path->completed_trainings = $completed;
                 $path->progress_percent = $trainingIds->count() > 0
-                    ? round(($completed / $trainingIds->count()) * 100)
+                    ? (int) round(($completed / $trainingIds->count()) * 100)
                     : 0;
                 return $path;
             });
